Extract Prompt Studio advanced controls
