+++ docs/development/issue-14-frontend-js-maintenance.md
@@ -197,6 +197,7 @@ web/js/
   easyuse_anima_prompt_studio.js
   prompt_studio/
     constants.js
+    advanced_controls.js
     utils.js
     settings.js
     text.js
@@ -228,26 +229,27 @@ same PR by request.
 Planned slices:
 
 1. `constants.js` and `utils.js`
-2. `settings.js`
-3. `text.js`
-4. `highlight.js`
-5. `schema.js`
-6. `state.js`
-7. `canvas_forwarding.js`
-8. `serialization.js`
-9. `layout.js`
-10. `textarea.js`
-11. `wheel.js`
-12. `dom.js`
-13. `extend_slots.js`
-14. `extend_slot_controls.js`
-15. `style.js`
-16. `tooltip.js`
-17. `widgets.js`
-18. `legend.js`
-19. `fields.js`
-20. `node_hooks.js`
-21. entry file slim-down
+2. `advanced_controls.js`
+3. `settings.js`
+4. `text.js`
+5. `highlight.js`
+6. `schema.js`
+7. `state.js`
+8. `canvas_forwarding.js`
+9. `serialization.js`
+10. `layout.js`
+11. `textarea.js`
+12. `wheel.js`
+13. `dom.js`
+14. `extend_slots.js`
+15. `extend_slot_controls.js`
+16. `style.js`
+17. `tooltip.js`
+18. `widgets.js`
+19. `legend.js`
+20. `fields.js`
+21. `node_hooks.js`
+22. entry file slim-down
 
 Each PR should mostly move code and update imports. Behavior changes should be
 separate PRs unless the move reveals a confirmed bug that cannot be separated.
@@ -714,6 +716,7 @@ Use this in Issue #14 or follow-up PR tracking comments.
 
 ### Phase 2: Prompt Studio module split
 [ ] constants.js extracted
+[ ] advanced_controls.js extracted
 [ ] utils.js extracted
 [ ] settings.js extracted
 [ ] text.js extracted
